feat(montgomery mul): Add some test case
